Created music builder
This is a simple HTML app which lets you put in the notes and durations. Then it generates the code that you would need to insert into the program to play the song specified utilizing the included buzzer.

If we choose a song, this system will make it easier to copy from sheet music to actual code.