Create bumperexample.ino
Here's the example file we created at the workshop today. None of the libraries are included... and this code doesn't work perfectly--there's some issue with the line following where it starts to hug the curve. 

This is related somehow to its sensor readings. I would put some sort of additional check in the sensor booleans, like "sensor_values[0] < QTR_THRESHOLD && currentDelay == 0" maybe. It's something to check out.

Make sure the threshold is okay.